Front matter mappings by category.
This is a little bit of a hack, but I want to be able to control how
front matter is set based on more than just the post type, and using
categories is useful.

The canned example now shows how a "tweet" category can be used to add
syndication front matter to a post.