Add support for image alt text
Image alt text is really important for people using screen readers, and
in the Mastodon world is an important part of the platform etiquette.

That alt text is passed along in an additional parameter in the request,
and so this change lifts that parameter out and makes it available for
population in the page front matter.