Fix many style violations
I wrote this when I was first learning Ruby and never bothered to use
Rubocop. Knowing what I know now, I've gone back and cleaned a few
things up. It's still a one-script tool which means it's a little
monolithic, but at least it's written in a way to (mostly) align with
modern style guides.